create LatestNotes.tsx
